backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
celery==5.3.4
neo4j==5.15.0
groq==0.4.1
sentence-transformers==2.7.0
transformers==4.36.0
huggingface-hub==0.19.4
torch==2.1.0
langchain==0.1.0
langchain-community==0.0.10
unstructured[pdf]==0.11.8
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
langdetect==1.0.9
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
aiofiles==23.2.1
httpx==0.25.2
numpy==1.24.3
scikit-learn==1.3.2